refactor(Register): move register to core directory
