+++ code/handlers/all/new_bing_handle.py
@@ -5,7 +5,7 @@ import json
 
 from EdgeGPT import Chatbot
 
-from handlers.register import msg_handle_register, bot_register
+from handlers import msg_handle_register, bot_register
 from core.data_structure import MsgInfo, ActionType
 from core.bot import bot_pool, BotBase
 from core.card_builder import CardBuilder, CardTemplate
